Create README.md
Abilities Validated by the Certification
Effectively demonstrate knowledge of how to architect and deploy secure and robust applications on AWS technologies
Define a solution using architectural design principles based on customer requirements
Provide implementation guidance based on best practices to the organization throughout the life cycle of the project
Recommended Knowledge and Experience
Hands-on experience using compute, networking, storage, and database AWS services
Hands-on experience with AWS deployment and management services
Ability to identify and define technical requirements for an AWS-based application
Ability to identify which AWS services meet a given technical requirement
Knowledge of recommended best practices for building secure and reliable applications on the AWS platform
An understanding of the basic architectural principles of building on the AWS Cloud
An understanding of the AWS global infrastructure
An understanding of network technologies as they relate to AWS
An understanding of security features and tools that AWS provides and how they relate to traditional services
Prepare for Your Exam
There is no better preparation than hands-on experience. There are many relevant AWS Training courses and other resources to assist you with acquiring additional knowledge and skills to prepare for certification. Please review the exam guide for information about the competencies assessed on the certification exam.